STR-492: Standard C99: report all warnings and standards compliance
The build system now provides warnings (and optionally errors) if the
code uses features that are not standard C99, or are at least in the
Single Unix Specification v3 (SUSv3).

For this, there is a new macro XT_STRICT_ISO_C99 that is always
invoked from XT_SETUP. If the compiler is gcc, then the variable
STR_CFLAGS (used by Makefile.xt) is set to:

  -std=c99 -Wall -pedantic -Wno-unused-label -Wno-unused-variable

The last two warnings are allowed because they occur frequently in
generated C code. We might remove these in the future.

Also, this macro adds an option --enable-werror, which makes the build
fail if there is an error. Be careful: this is propagated to the
installed Stratego compiler.

In the Stratego libraries, we now need a define to allow C functions
that are not specified by C99. After a few experiments and browsing
for bug reports on the web, it seems that the best option is to allow
all C functions from the SUSv3, which can be achieved using the define
_XOPEN_SOURCE=600. This define has selectively been added to makefiles
that compile hand-written native code.

Unfortunately, we use a few functions that are not in POSIX.1-2001, so
we had to fall back to the entire SUS. Where possible, I've used the
define -D_POSIX_C_SOURCE=200112L for POSIX.1-2001. Maybe we should try
to achieve this everywhere.

----------------------------------------------------------------
Fixes:

* autoxt/autoxt.m4

Introduced the XT_STRICT_ISO_C99 macro.

* stratego-libraries/lib/native/stratego-lib/system-io-dir.c

Dropped the setenv fallback. Setenv is in SUSv3, so there is no reason
for falling back to putenv. Putenv itself was causing some problems as
well: it is not POSIX.

* stratego-libraries/lib/native/stratego-lib/system-posix-file.c

Dropped the optmized implementation of mkdtemp, since it will never be
defined anymore. In the GNU C library it is only enabled on __USE_BSD
(which is defined if there is a _GNU_SOURCE define). With our stricter
check for standard compliance, mkdtemp would never be defined anymore,
which makes the if useless. The consequence is that mkdtemp is now
always unsafe in Stratego.

* stratego-libraries/build-tools/cgen-aterm.str:

don't use const strings: the aterm library does not feature const
indicators, which causes warnings, which cause errors if you enable
-Werror.

* stratego-libraries/sglr/justsglr/Makefile.am

Removed $(STR_CFLAGS): sglr is not C99 compatible and will soon be
moved out of Stratego/XT now there is a new sdf-bundle that uses
libtool ot create dynamic libraries.

* stratego-libraries/sglr/lib/Makefile.am

Enable POSIX functions

* stratego-libraries/lib/spec/Makefile.am
* stratego-libraries/lib/native/stratego-lib/Makefile.am

Enable SUSv3 functions

* stratego-libraries/configure.ac

Dumped the obsolete checks for getpagesize and mkdtemp

svn path=/strategoxt/trunk/; revision=15523