STR-311
Introduced a new compiler component 'frontend', which combines a large
part of the frontend components (it is not yet activated in the
compiler, nor tested). This should reduce the time spent in the
frontend of the compiler by avoiding intermediate files, and is the
first step to refactoring the compiler to apply a longer pipeline of
transformations to a single definition. 

The frontend does not yet call the format checker, which would require
inlining of the RTG for the various formats.

Extract was broken due to dynamic rules (names) that it had in common
with other components. Fixed by renaming the dynamic rules.


svn path=/strategoxt/trunk/; revision=15541