STR-87:  Separate match, build and congruence term patterns in syntax definition. STR-600: Poor error report for 'build default' in normal pattern STR-576: wildcards in build are not caught by strc STR-606: illegal constructs in builds
All these issues were concerned with the same problem; match symbols
in build patterns caused the compiler to fail in the back-end without
proper explanation. 

The front-end now includes a test that forbids wildcard and @ (as) in
builds. The check also forbids build defaults (_ t) in match *and*
build patterns. (Build defaults should only be used in overlays, and
are eliminated upon overlay expansion.) Wildcards and @ are also
forbidden in overlay definitions.

    strc-core/lib/strc.str
    strc-core/lib/stratego/strc/model/model.str
    strc-core/lib/stratego/strc/front/frontend.str

A bunch of tests for the issue:

    strc-core/spec/test1/Expressions.str
    strc-core/spec/test1/test106.str
    strc-core/spec/test1/Makefile.am
    strc-core/spec/testneg/test09.str
    strc-core/spec/testneg/test10.str
    strc-core/spec/testneg/Makefile.am

Use pre-compiled library in these tests to speed up compilation.

    strc-core/spec/testneg/test04.str
    strc-core/spec/testneg/test05.str
    strc-core/spec/testneg/test08.str

STR-636 Refactoring: Strategies for checking definedness of calls and
removing unused calls in separate module.

    strc-core/lib/stratego/strc/front/call-graph.str

Commented this test; it is for contextual rules, which are obsolete.

    strc-core/spec/test1/test44.str

svn path=/strategoxt/trunk/; revision=15784