The XT_USE_BOOTSTRAP_XT_PACKAGES macro is now a bit more clever:
if you specify --with-strategoxt=.., then this implies --enable-bootstrap

if you specify --with-xtc=..., then this implies --enable-xtc

The --enable-bootstrap and --enable-xtc options are still useful for
the cases where the user is not interested in specifying the location
of packages.

New automake conditional: XT_BOOTSTRAP, which can be used to disable
stuff that is only working when the full Stratego/XT is available. An
example of this is the testing of a grammar using parse-unit.


svn path=/strategoxt/trunk/; revision=15859