STR-646 Hashtable folds
Applying operations to the elements of a hashtable currently requires
the creation of a list of keys. If one is only interested in
operations with side-effects on the elements (e.g., changing the
entry, creating dynamic rules, printing) the construction of the list
is wasteful, especially when it concerns large tables. A fold
operation on the keys/values of a table implemented using iteration
rather than recursion, allows one to avoid list construction, while
still providing the possibility of building up a composite result from
the entries.

The following library strategies implement these operations:

  hashtable-fold-keys(s | acc)
  hashtable-fold-values(s | acc)

The parameter strategy s should be callable as s(|acc), and apply to
the key or value, returning a new accumlator value.  

STR-647 Iterative list loop and fold  

Provide iterative loops to visit or fold the elements of a list
avoiding recursion, and reconstruction of the list as in the case of
map. (Based on a prototype by Martin Bravenboer) This is implemented
by the following library strategies:

  list-loop(s) 
    applies s to each element of the list returning the
    original list

  list-fold(s | acc)
    folds the list using accumulator acc, i.e., a list [t1,...,tn]
    is conceptually folded as <s(|<s(|<s(|acc)> t1)> t2 ...)> tn, but
    this is implemented using iteration rather than recursion

(The compiler should support a more principled approach for turning
recursion into iteration; tail recursion elimination!)


svn path=/strategoxt/trunk/; revision=15880