M    autoxt/Makefile.xt
Temporary hack to fix a build problem with the new aterm library 2.5,
where ATfgetLength is a function (used to be a macro).

Warning: the next baseline is a hack and will only work at Linux. The
baseline after this will fix the problem.

M    stratego-libraries/lib/native/stratego-lib/stratego-lib-common.c

_ATerm is a union in aterm library 2.5, not a struct.


svn path=/strategoxt/trunk/; revision=16202