M    stratego-regular/syn/rg-term.sdf
STR-676: added syntactic support for <string> and <int> terms. Note
that this is not yet supported by format-check tools, and sdf2rtg does
not generate the new construct yet.

M    stratego-regular/syn/rg-nonterm.sdf

Added syntax for sets non-terminals (which has no semantic meaning
whatsoever: it is still a single non-terminals).

This is useful/necessary for the upcoming format-checker, which uses a
deterministic regular tree automaton. To get a deterministic tree
automaton, states are introduced for sets of non-terminals. These sets
are now supported syntactically, which is mostly useful for the
development of the new format checker (STR-398).


svn path=/strategoxt/trunk/; revision=16268