STR-577: support import of an ATerm from a file
Added support for bigger aterms (beyond the limit of C line size and
Stratego list processing).

M    strc-core/lib/stratego/strc/c/s2c-ng.str
A    strc-core/lib/libstrc-native.c

Use native code to construct the C array. Stratego cannot construct an
array init for large aterms, which results in lists of chars of more
than 500.000 elements (stack overflow). Also, the pretty-printer
horribly fails.

Also, it is not possible to instruct the pretty-printer to break the
array init over multiple lines.

The solution is a very dirty hack: just include the entire array init
as a string in the value of a single integer constant. Thus, the AST
and pretty-printer think that there is only a single expression, but
actually this will print the entire array init. Of course, a
pretty-printer should never allow this, but in this case it is
fortunate that this works.

I used bin2c from Nix (thanks to Eelco Dolstra) to get all the
unsigned/signed stuff right.

This work also revealed a bug in the aterm shared textual format,
which cannot print an empty tuple () and in some cases even produces
an incorrect aterm. This issue has been reported:

  http://sjofar.sen.cwi.nl:8080/show_bug.cgi?id=677

I returned to the use of binary aterms (an intermediate solution was
using shared textual aterms, which is a bit easier to get platform
independent). The loading a aterms is now even faster than it used to
be with the cgen-aterm tool. Also, it is faster than reading from a
file.

AM   strc-core/spec/test1/Java-15.tbl
A    strc-core/spec/test1/test107.str

A very big test (>600.000 chars)


svn path=/strategoxt/trunk/; revision=16335