M    stratego-front/syn/StrategoStratego.sdf A    stratego-front/syn/EmbeddedStratego.sdf
Upgraded to generic embedding module 

StrategoStratego is now defined as:

  ------------------------------
  module StrategoStratego
  imports
    StrategoMix[StrategoHost]
    EmbeddedStrategoMix[StrategoObject PreTerm[[StrategoHost]]]

  hiddens
    context-free start-symbols Module[[StrategoHost]]
  ------------------------------

EmbeddedStrategoMix can also be used in legacy grammars, where the
prefixed version of Stratego is used. One example is
Stratego-Stratego-C, which is now defined as:

  ------------------------------
  module Stratego-Stratego-C
  imports
    Stratego-C
    StrategoRenamed
    EmbeddedStrategoMix[StrategoObject StrategoPreTerm]

  hiddens
    context-free start-symbols StrategoModule
  ------------------------------

The reason for using StrategoRenamed is that Stratego-C is not yet
updated, so it is impossible to avoid the use of a prefixed host
language. In the end, all this stuff should be updated, but most
embeddings don't have a testsuite, so it is rather easy to break them.


svn path=/strategoxt/trunk/; revision=16424