M    stratego-front/test/syn/StrategoStratego/preterm.testsuite
Test a few more embedding things

M    stratego-front/syn/EmbeddedStratego.sdf
M    stratego-front/syn/StrategoStratego.sdf

EmbeddedStratego now has two parameters: one for the quotations and
one for the anti-quotations. This is necessary for STR-685, but is
also useful in general: for languages that encode priorities in
expression non-terminals, different non-terminals might be necessary
for the anti-quotations and quotations.

A    stratego-front/test/anti-quote-preterm.str

Test for problem that popped up in Transformers code.

/**
 * (STR-685) meta-explode does not support PreTerm anti-quotations
 *
 * If an anti-quotation is defined as a PreTerm, then
 * this example fails. The problem is that a meta-explode
 * assumes that FromTerm has a Term as an argument. Often,
 * this indeed accidentally the case, since most non-constant
 * PreTerms are also Terms. However, if constant terms
 * are anti-quoted, then compilation will fail.
 *
 * This test checks two different things:
 * - if StrategoStratego uses Term anti-quotation
 * - if StrategoStratego uses PreTerm, then meta-explode is checked
 */


svn path=/strategoxt/trunk/; revision=16440