Dropped use of autoheader in stratego-libraries.
stratego-config.h is no longer installed in the includedir of
strategoxt. There is no configuration option in this file that is
still necessary as part of the strategoxt installation. The mechanism
of config.h installation is rather fragile (some defines have to be
removed), so if it's not necessary anymore, then why not drop it ...

Another reason for this change is that several C files did not
correctly include the generated config.h before aterm2.h. This breaks
on x86-64 systems (STR-698) because aterm2.h requires defines of
various SIZEOF_ variables that now got defined only after the
inclusion of aterm2.h.

This change might break libstratego-sglr on some platforms: I vaguely
recall some problems with compilation for C99, POSIX, and POSIX+XSI
that were only solved by *not* including config.h. However, if this is
still the case, then it just needs to be solved in a more solid and
explicit way.


svn path=/strategoxt/trunk/; revision=16735