M    sdf-tools/analyse/parse-unit/parse-unit-present.str
Print the expected and actual outputs of parsing using
write-in-text-to-stream, not using debug. Debug prints aterm strings
without double quotes, which is confusing if the actual output of
parsing is a string as well (uncommon, but well ... )

M    stratego-libraries/aterm/lib/stratego/aterm/Explode.str

Added support for imploding and exploding aterms with annotations. The
implode-aterm and explode-aterm strategies are mostly used by
meta-tools (such as parse-unit) for comparing aterms.

This fixes (STR-708) parse-unit: support annotations in expected and
actual results.

A    stratego-libraries/aterm/tests/explode-implode-tests.str

Started a testsuite for implode/explode-aterm. Currently the testsuite
performs some roundtripping tests.


svn path=/strategoxt/trunk/; revision=16880