M    strc-core/lib/stratego/strc/c/s2c-ng.str
CallDynamic: gracefully handle a strategy that is not defined in the
table for dynamic calls. The current behaviour was a SEGV.

Maybe we should actually print some warning as well?

M    stratego-libraries/runtime/lib/stratego-dynamic-call.c

SRTS_lookup_function is not a normal strategy: it returns a Strategy
closure. Therefore, it makes no sense to invoke _fail if the strategy
is not available in the table. Changed the return value to NULL (which
is actually the same as _fail, so this is just for clarity).


svn path=/strategoxt/trunk/; revision=17625