For each dynamic rules DR generate a derived rule
  fold-DR(f,...)

which takes a folding strategy as first parameter and
the regular parameters of DR as other parameters.

This can be used as an alternative to chain-DR,  providing
more control over the folding process. One example usage
is:

  let f(s) = foldl(s <+ Snd) in fold-DR(f) end

This application ignores a rule application (s) if it fails.

M    strc-core/lib/stratego/strc/front/lift-dynamic-rules.str

svn path=/strategoxt/trunk/; revision=17678