Fix ambiguity in XML syntax + removed some unused productions.
The ambiguity is that, when more than one consecutive CharDataPart is present,
the CharDataParts aren't put greedily into a *single* CharData.
(And multiple consecutive CharData is valid too...)

In the C tooling this ambiguity was eaten by implode-asfix. (though visible when using sglri)
In JSGLR, it made it impossible to parse certain XML files of a few KB only within a finite amount of time and memory.
(More precisely, recognizing was fine, but at least one of tree building / disambiguation / tree outputting was impossible.)

svn path=/strategoxt/trunk/; revision=23117