Revert "Remove -Wno-unused-but-set-variable flag"
This reverts commit 8821a02ab7860df6f47792a04ab181bcecdb0842.

The -Wno-unused... was required to have the build succeed. The generated C code has variables that are set but unread. Because configure is run with -Werror the otherwise warnings get elevated to errors causing the build to fail. The -Wno-unused... flag lowered those back to warnings.

The issue remains that the -Wno-unused-... flag is not supported on LLVM gcc.