imploder bugfix: create more sensible Conc constructors.
This is for Cons/Snoc cases, where the single elem is now wrapped in a list.
The change is necessary for the new parsetable generator which uses Snoc lists instead or Conc lists, which means the Conc looks different when you use list meta-variables or FromTerm instead of a sublist (as is used in Stratego mix syntax aka concrete object syntax).